Refactor MainScreen: Extract Sidebar and SidebarToggle components
