+++ app/components/index.ts
@@ -1,0 +1,10 @@
+// UI Components
+export { Sidebar } from './Sidebar';
+export { SidebarToggle } from './SidebarToggle';
+export { Collapsible } from './Collapsible';
+export { ExternalLink } from './ExternalLink';
+export { HapticTab } from './HapticTab';
+export { HelloWave } from './HelloWave';
+export { ParallaxScrollView } from './ParallaxScrollView';
+export { ThemedText } from './ThemedText';
+export { ThemedView } from './ThemedView';
